+++ index.js
@@ -1,54 +1,15 @@
-/* Styles to be used by the other builder class */
-const styles = {
-  fgColors: {
-    black:   [30, 39],
-    red:     [31, 39],
-    green:   [32, 39],
-    yellow:  [33, 39],
-    blue:    [34, 39],
-    magenta: [35, 39],
-    cyan:    [36, 39],
-    white:   [37, 39],
-    gray:    [90, 39]
-  },
-  bgColors: {
-    bgBlack:   [40, 49],
-    bgRed:     [41, 49],
-    bgGreen:   [42, 49],
-    bgYellow:  [43, 49],
-    bgBlue:    [44, 49],
-    bgMagenta: [45, 49],
-    bgCyan:    [46, 49],
-    bgWhite:   [47, 49],
-    bgGray:    [100, 49]
-  },
-  modifiers: {
-    Reset: "\x1b[0m",
-    Bold: "\x1b[1m",
-    Dim: "\x1b[2m",
-    Underscore: "\x1b[4m",
-    Blink: "\x1b[5m",
-    Reverse: "\x1b[7m",
-    Hidden: "\x1b[8m"
-  }
-}
+import {bgRed, blue, green, yellow, red, black, white, magenta} from "./dye.js";
+import dye from './dye.js';
 
-function buildStyles() {
-  for (const [groupname, group] of Object.entries(styles)) {
-    for (const [stylename, style] of Object.entries(group)) {
-      styles[stylename] = {
-        open: `\x1b[${style[0]}m`,
-        close: `\x1b[${style[1]}m`,
-      };
-    }
-  }
+console.log(blue("Hello, World!"));
+console.log(green("Hello, World!"));
+console.log(yellow("Hello, World!"));
+console.log(red("Hello, World!"));
+console.log(white("Hello, World!"));
+console.log(magenta("Hello, World!"));
+console.log(black("Hello, World!"));
 
-  styles.fgColors.close = '\x1b[39m';
-  styles.bgColors.close = '\x1b[49m';
 
-  return styles;
-}
+console.log(bgRed(blue("Hello, World!")));
 
-const ansistyles = buildStyles();
-
-export default ansistyles;
+//console.log(dye.blue('Hello, World!'));
